Fixed some errors and integrated with extracted out TaskBuilder from event branch
